+++ app/Http/Controllers/Admin/Users/UsersController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin\Users;
 
 use App\Http\Controllers\Controller;
 use App\Models\Goal\Goal;
+use App\Models\SetResidencial\Setresidencial;
 use Illuminate\Http\Request;
 use App\Models\State\State;
 use App\Models\User;
@@ -11,6 +12,7 @@ use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\Rule;
 use Spatie\Permission\Models\Role;
 use Illuminate\Database\QueryException;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\View;
 
 
@@ -25,22 +27,57 @@ class UsersController extends Controller
     }
     public function index()
     {
-        $users = User::all();
-        return view('admin.users.index', compact('users'));
+        if(auth()->user()->state_id == 2){
+            Auth::logout();
+            return redirect()->route('login')->with('info', 'EL USUARIO SE ENCUENTRA EN ESTADO INACTIVO EN EL SISTEMA POR FAVOR CONTACTAR A UN ADMINISTRADOR.');
+        }
+
+        if (auth()->user()->hasRole('ADMINISTRADOR')) {
+            $users = User::all();
+            return view('admin.users.index', compact('users'));
+        }elseif (auth()->user()->hasRole('SUB_ADMINISTRADOR')) {
+
+            $setresidencialIds = auth()->user()->setresidencials->pluck('id')->toArray();
+            $users = User::whereHas('setresidencials', function ($query) use ($setresidencialIds) {
+                $query->whereIn('setresidencial_id', $setresidencialIds);
+            })
+            ->whereHas('setresidencials') // Aseguramos que los usuarios tengan al menos un conjunto residencial
+            ->get();
+
+            return view('admin.users.index', compact('users'));
+        }
     }
 
     public function create()
     {
-        $users = User::all();
-        $states = State::all();
-        $roles = Role::all();
-        $goals = Goal::where('state_id', 1)->get();
+        if(auth()->user()->state_id == 2){
+            Auth::logout();
+            return redirect()->route('login')->with('info', 'EL USUARIO SE ENCUENTRA EN ESTADO INACTIVO EN EL SISTEMA POR FAVOR CONTACTAR A UN ADMINISTRADOR.');
+        }
 
-        return view('admin.users.create', compact('users', 'states', 'roles','goals'));
+        if (auth()->user()->hasRole('ADMINISTRADOR')) {
+            $states = State::all();
+            $roles = Role::all();
+            $goals = Goal::where('state_id', 1)->get();
+            $setresidencials = Setresidencial::where('state_id', 1)->get();
+            return view('admin.users.create', compact( 'states', 'roles','goals','setresidencials'));
+        }elseif (auth()->user()->hasRole('SUB_ADMINISTRADOR')) {
+            $setresidencialIds = auth()->user()->setresidencials->pluck('id')->toArray();
+            $states = State::all();
+            $roles = Role::all();
+            $goals = Goal::whereIn('setresidencial_id', $setresidencialIds)->where('state_id', 1)->get();
+            $setresidencials = auth()->user()->setresidencials()->where('state_id', 1)->get();
+            return view('admin.users.create', compact( 'states', 'roles','goals','setresidencials'));
+        }
     }
 
     public function store(Request $request)
     {
+        if(auth()->user()->state_id == 2){
+            Auth::logout();
+            return redirect()->route('login')->with('info', 'EL USUARIO SE ENCUENTRA EN ESTADO INACTIVO EN EL SISTEMA POR FAVOR CONTACTAR A UN ADMINISTRADOR.');
+        }
+
         $request->validate([
             'name' => 'required',
             'lastname' => 'required',
@@ -51,6 +88,7 @@ class UsersController extends Controller
             'state_id' => 'required',
             'roles' => ['required', 'array', 'min:1'],
             'goals' => ['array', 'exists:goals,id'],
+            'setresidencials' => ['array', 'exists:setresidencials,id'],
         ]);
         $user = User::create([
             'name' => $request->name,
@@ -64,38 +102,95 @@ class UsersController extends Controller
     
         $user->roles()->sync($request->roles);
         $user->goals()->sync($request->goals); 
+        $user->setresidencials()->sync($request->setresidencials); 
 
         return redirect()->route('admin.users.index')->with('success', 'EL USUARIO SE CREO CORRECTAMENTE.');
     }
 
     public function show(User $user)
     {
+        if(auth()->user()->state_id == 2){
+            Auth::logout();
+            return redirect()->route('login')->with('info', 'EL USUARIO SE ENCUENTRA EN ESTADO INACTIVO EN EL SISTEMA POR FAVOR CONTACTAR A UN ADMINISTRADOR.');
+        }
+
         return view('admin.users.show', compact('user'));
     }
 
     public function edit(User $user)
     {
-        $states = State::all();
-        $roles = Role::all();
-        $goals = Goal::where('state_id', 1)->get();
+        if (auth()->user()->state_id == 2) {
+            Auth::logout();
+            return redirect()->route('login')->with('info', 'EL USUARIO SE ENCUENTRA EN ESTADO INACTIVO EN EL SISTEMA POR FAVOR CONTACTAR A UN ADMINISTRADOR.');
+        }
     
-        // Obtener los IDs de las metas ya asignadas al usuario
-        $goals_user = $user->goals->pluck('id')->toArray();
+        if (auth()->user()->hasRole('ADMINISTRADOR')) {
+            $states = State::all();
+            $roles = Role::all();
+            
+            $goals = Goal::where('state_id', 1)
+                ->orWhere(function ($query) use ($user) {
+                    $query->where('state_id', 2)
+                          ->whereHas('users', function ($q) use ($user) {
+                              $q->where('user_id', $user->id);
+                          });
+                })
+                ->get();
     
-        return view('admin.users.edit', compact('user', 'states', 'roles', 'goals', 'goals_user'));
+            $setresidencials = Setresidencial::where('state_id', 1)
+                ->orWhere(function ($query) use ($user) {
+                    $query->where('state_id', 2)
+                        ->whereHas('users', function ($q) use ($user) {
+                            $q->where('user_id', $user->id);
+                        });
+                })
+            ->get();
+            $goals_user = $user->goals->pluck('id')->toArray();
+            $setresidencials_user = $user->setresidencials->pluck('id')->toArray();
+        
+            return view('admin.users.edit', compact('user', 'states', 'roles', 'goals', 'setresidencials', 'goals_user', 'setresidencials_user'));
+        } elseif (auth()->user()->hasRole('SUB_ADMINISTRADOR')) {
+            $setresidencialIds = auth()->user()->setresidencials->pluck('id')->toArray();
+        
+            $states = State::all();
+            $roles = Role::all();
+            
+            $goals = Goal::whereIn('setresidencial_id', $setresidencialIds)
+                ->where('state_id', 1)
+                ->orWhere(function ($query) use ($user) {
+                    $query->where('state_id', 2)
+                          ->whereHas('users', function ($q) use ($user) {
+                              $q->where('user_id', $user->id);
+                          });
+                })
+                ->get();
+    
+            $setresidencials = auth()->user()->setresidencials()->where('state_id', 1)->get();
+            $goals_user = $user->goals->pluck('id')->toArray();
+            $setresidencials_user = $user->setresidencials->pluck('id')->toArray();
+        
+            return view('admin.users.edit', compact('user', 'states', 'roles', 'goals', 'setresidencials', 'goals_user', 'setresidencials_user'));
+        }
     }
+    
     
 
 
     public function update(Request $request, User $user)
     {
+        if(auth()->user()->state_id == 2){
+            Auth::logout();
+            return redirect()->route('login')->with('info', 'EL USUARIO SE ENCUENTRA EN ESTADO INACTIVO EN EL SISTEMA POR FAVOR CONTACTAR A UN ADMINISTRADOR.');
+        }
+
         $request->validate([
             'name' => 'required',
             'lastname' => 'nullable',
             'email' => ['required', 'email'],
             'state_id' => 'required',
             'roles' => ['required', 'array', 'min:1'],
-            'goals' => ['array', 'exists:goals,id'], // Validar metas seleccionadas
+            'goals' => ['array', 'exists:goals,id'], 
+            'setresidencials' => ['array', 'exists:setresidencials,id'], 
         ]);
 
         $data = $request->all();
@@ -108,7 +203,8 @@ class UsersController extends Controller
 
         $user->update($data);
         $user->roles()->sync($request->roles);
-        $user->goals()->sync($request->goals); // Sincronizar metas
+        $user->goals()->sync($request->goals); 
+        $user->setresidencials()->sync($request->setresidencials); 
 
         return redirect()->route('admin.users.index')->with('edit', 'EL USUARIO SE EDITÓ CORRECTAMENTE.');
     }
@@ -116,17 +212,27 @@ class UsersController extends Controller
 
     public function destroy(User $user)
     {
+        if(auth()->user()->state_id == 2){
+            Auth::logout();
+            return redirect()->route('login')->with('info', 'EL USUARIO SE ENCUENTRA EN ESTADO INACTIVO EN EL SISTEMA POR FAVOR CONTACTAR A UN ADMINISTRADOR.');
+        }
+
         if ($user->id === 1) {
             return redirect()->route('admin.users.index')->with('info', 'ESTE USUARIO NO SE PUEDE ELIMINAR YA QUE ES UNO DE LOS PRINCIPALES EN EL SISTEMA');
         }
 
         try {
+            $user->roles()->detach();
+            $user->goals()->detach();
+            $user->setresidencials()->detach();  
+
             $user->delete();
             return redirect()->route('admin.users.index')->with('delete', 'EL USUARIO SE ELIMINO CORRECTAMENTE.');
         } catch (QueryException $e) {
             if ($e->errorInfo[1] === 1451) {
                 return redirect()->route('admin.users.index')->with('info', 'EL USUARIO NO SE PUEDE ELIMINAR YA QUE ESTA RELACIONADO CON OTRO REGISTRO.');
             }
+            return redirect()->route('admin.users.index')->with('info', 'OCURRIÓ UN ERROR AL INTENTAR ELIMINAR EL USUARIO.');
         }
     }
 }
